rename services package to service
