+++ service/github.go
@@ -1,4 +1,4 @@
-package services
+package service
 
 import (
 	"context"
